Refactor to avoid package naming collision from client capauth.
